# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: POS Platform API
  version: 0.1.0
  description: >
    REST API for the POS platform providing checkout, backoffice, and
    synchronization capabilities. The API is designed to be idempotent, secure,
    and ready for hybrid deployments.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/admin/backups:
    get:
      tags:
        - Administration
      summary: List backups
      responses:
        '200':
          description: Available system backups.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SystemBackup'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/admin/backup:
    post:
      tags:
        - Administration
      summary: Create backup
      description: Triggers a full database backup and registers the archive.
      responses:
        '201':
          description: Backup created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemBackup'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/admin/backup/upload:
    post:
      tags:
        - Administration
      summary: Upload backup archive
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Backup archive (.zip)
              required:
                - file
      responses:
        '201':
          description: Backup uploaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemBackup'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/admin/backup/{filename}:
    delete:
      tags:
        - Administration
      summary: Delete backup
      parameters:
        - $ref: '#/components/parameters/BackupFilename'
      responses:
        '204':
          description: Backup removed.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/admin/backup/{filename}/download:
    get:
      tags:
        - Administration
      summary: Download backup archive
      parameters:
        - $ref: '#/components/parameters/BackupFilename'
      responses:
        '200':
          description: Backup archive.
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/admin/restore:
    post:
      tags:
        - Administration
      summary: Restore backup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
      responses:
        '200':
          description: Restore accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestoreResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/admin/import/legacy:
    post:
      tags:
        - Administration
      summary: Import legacy DBF data
      description: >
        Uploads legacy DBF/DBT files (or a ZIP archive) and queues the ETL pipeline
        to rebuild the normalized database from legacy staging tables.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LegacyImportRequest'
      responses:
        '202':
          description: Import accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegacyImportResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/admin/import/legacy/{sessionId}/status:
    get:
      tags:
        - Administration
      summary: Get legacy import status
      parameters:
        - $ref: '#/components/parameters/ImportSessionId'
      responses:
        '200':
          description: Current status for the legacy import.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegacyImportStatus'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/admin/import/legacy/{sessionId}/report:
    get:
      tags:
        - Administration
      summary: Download legacy import reconciliation report
      parameters:
        - $ref: '#/components/parameters/ImportSessionId'
      responses:
        '200':
          description: CSV reconciliation report.
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
components:
  parameters:
...
    BackupFilename:
      name: filename
      in: path
      required: true
      description: Backup archive filename.
      schema:
        type: string
    ImportSessionId:
      name: sessionId
      in: path
      required: true
      description: Import session identifier.
      schema:
        type: string
...
  responses:
...
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: resource not found
...
    UnauthorizedError:
      description: Authentication credentials are missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: unauthorized
    ForbiddenError:
      description: Authenticated user lacks permission for the requested action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: forbidden
  schemas:
...
    SystemBackup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        sizeBytes:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        checksum:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        createdBy:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            fullName:
              type: string
          required:
            - id
            - fullName
      required:
        - id
        - filename
        - sizeBytes
        - createdAt
    RestoreRequest:
      type: object
      properties:
        file:
          type: string
        password:
          type: string
        confirm:
          type: boolean
          description: Confirmation flag. Must be true.
      required:
        - file
        - password
        - confirm
    RestoreResponse:
      type: object
      properties:
        status:
          type: string
        restored:
          type: boolean
      required:
        - status
        - restored
    LegacyImportRequest:
      type: object
      properties:
        files:
          type: array
          items:
            type: string
            format: binary
        overwrite:
          type: boolean
          default: false
        password:
          type: string
        confirmation:
          type: string
      required:
        - files
        - password
        - confirmation
    LegacyImportResponse:
      type: object
      properties:
        status:
          type: string
        sessionId:
          type: string
        importId:
          type: string
          format: uuid
        overwrite:
          type: boolean
        files:
          type: array
          items:
            type: string
        message:
          type: string
      required:
        - status
        - sessionId
        - overwrite
        - files
        - message
    LegacyImportLogEntry:
      type: object
      properties:
        level:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - level
        - message
        - createdAt
    LegacyImportStatus:
      type: object
      properties:
        status:
          type: string
        overwrite:
          type: boolean
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        summary:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
        reportAvailable:
          type: boolean
        logs:
          type: array
          items:
            $ref: '#/components/schemas/LegacyImportLogEntry'
      required:
        - status
        - overwrite
        - createdAt
        - reportAvailable
        - logs
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          nullable: true
      required:
        - message
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT